move `compute_output_shape` to `utils.py`
