+++ pyproject.toml
@@ -87,6 +87,7 @@ ignore = [
   "PLR",  # Pylint Refactor
   "PTH123", # `open()` should be replaced by `Path.open()`
   "PTH207", # "Replace `glob` with `Path.glob` or `Path.rglob`
+  "ISC001", # The following rules may cause conflicts when used with the formatter
 ]
 
 exclude = [
