+++ src/dolphin/__init__.py
@@ -4,5 +4,5 @@ from dolphin._version import version as __version__
 
 from ._log import *
 from ._blocks import *
-from ._readers import *
 from ._show_versions import *
+from .io import *
