+++ src/dolphin/phase_link/covariance.py
@@ -14,9 +14,8 @@ import numpy as np
 import pymp
 from numba import cuda, njit
 
-from dolphin._blocks import compute_out_shape
 from dolphin._types import Filename
-from dolphin.utils import _get_slices
+from dolphin.utils import _get_slices, compute_out_shape
 
 # CPU version of the covariance matrix computation
 _get_slices = njit(_get_slices)
