+++ src/dolphin/shp/_kld.py
@@ -6,7 +6,7 @@ import numba
 import numpy as np
 from numpy.typing import ArrayLike
 
-from dolphin._blocks import compute_out_shape
+from dolphin.utils import compute_out_shape
 
 from ._common import _make_loop_function, _read_cutoff_csv
