+++ src/dolphin/shp/_ks.py
@@ -8,9 +8,8 @@ import numpy as np
 from numba import cuda
 from numpy.typing import ArrayLike
 
-from dolphin._blocks import compute_out_shape
 from dolphin._log import get_log
-from dolphin.utils import _get_slices
+from dolphin.utils import _get_slices, compute_out_shape
 
 from ._common import remove_unconnected
